+++ tests/evaluate/test_default_evaluator.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 import io
 import json
-import shutil
-import tempfile
 from os.path import join as path_join
 from pathlib import Path
 from tempfile import TemporaryDirectory
@@ -15,7 +13,6 @@ import numpy as np
 import pandas as pd
 import pytest
 from PIL import Image, ImageChops
-from pyspark.sql import SparkSession
 from sklearn.datasets import load_breast_cancer, load_iris
 from sklearn.linear_model import LinearRegression, LogisticRegression
 from sklearn.metrics import f1_score, precision_score, recall_score
@@ -3626,132 +3623,3 @@ def test_evaluate_custom_metric_with_string_type():
             data["text"],
             check_names=False,
         )
-
-
-def test_write_to_delta_fails_without_spark():
-    with mlflow.start_run():
-        model_info = mlflow.pyfunc.log_model(
-            artifact_path="model", python_model=language_model, input_example=["a", "b"]
-        )
-        data = pd.DataFrame({"text": ["Hello world", "My name is MLflow"]})
-        with pytest.raises(
-            MlflowException,
-            match="eval_results_path is only supported in Spark environment",
-        ):
-            mlflow.evaluate(
-                model_info.model_uri,
-                data,
-                extra_metrics=[mlflow.metrics.latency()],
-                evaluators="default",
-                evaluator_config={
-                    "eval_results_path": "my_path",
-                    "eval_results_mode": "overwrite",
-                },
-            )
-
-
-@pytest.fixture
-def spark_session_with_tempdir():
-    tmpdir = tempfile.mkdtemp()
-    spark = (
-        SparkSession.builder.master("local[*]")
-        .config("spark.jars.packages", "io.delta:delta-spark_2.12:3.0.0")
-        .config("spark.sql.extensions", "io.delta.sql.DeltaSparkSessionExtension")
-        .config(
-            "spark.sql.catalog.spark_catalog", "org.apache.spark.sql.delta.catalog.DeltaCatalog"
-        )
-        .config("spark.sql.warehouse.dir", tmpdir)
-        .getOrCreate()
-    )
-
-    yield spark, tmpdir
-
-    spark.stop()
-    shutil.rmtree(tmpdir)
-
-
-def test_write_to_delta_fails_with_invalid_mode(spark_session_with_tempdir):
-    with mlflow.start_run():
-        model_info = mlflow.pyfunc.log_model(
-            artifact_path="model", python_model=language_model, input_example=["a", "b"]
-        )
-        data = pd.DataFrame({"text": ["Hello world", "My name is MLflow"]})
-        with pytest.raises(
-            MlflowException,
-            match="eval_results_mode can only be 'overwrite' or 'append'",
-        ):
-            mlflow.evaluate(
-                model_info.model_uri,
-                data,
-                extra_metrics=[mlflow.metrics.latency()],
-                evaluators="default",
-                evaluator_config={
-                    "eval_results_path": "my_path",
-                    "eval_results_mode": "invalid_mode",
-                },
-            )
-
-
-def test_write_eval_table_to_delta(spark_session_with_tempdir):
-    spark_session, tmpdir = spark_session_with_tempdir
-    with mlflow.start_run():
-        model_info = mlflow.pyfunc.log_model(
-            artifact_path="model", python_model=language_model, input_example=["a", "b"]
-        )
-        data = pd.DataFrame({"text": ["Hello world", "My name is MLflow"]})
-        results = mlflow.evaluate(
-            model_info.model_uri,
-            data,
-            extra_metrics=[mlflow.metrics.latency()],
-            evaluators="default",
-            evaluator_config={
-                "eval_results_path": "my_path",
-                "eval_results_mode": "overwrite",
-            },
-        )
-
-        eval_table = results.tables["eval_results_table"].sort_values("text").reset_index(drop=True)
-
-        eval_table_from_delta = (
-            spark_session.read.format("delta")
-            .load(f"{tmpdir}/my_path")
-            .toPandas()
-            .sort_values("text")
-            .reset_index(drop=True)
-        )
-
-        pd.testing.assert_frame_equal(eval_table_from_delta, eval_table)
-
-
-def test_write_eval_table_to_delta_append(spark_session_with_tempdir):
-    spark_session, tmpdir = spark_session_with_tempdir
-    with mlflow.start_run():
-        model_info = mlflow.pyfunc.log_model(
-            artifact_path="model", python_model=language_model, input_example=["a", "b"]
-        )
-        data = pd.DataFrame({"text": ["Hello world", "My name is MLflow"]})
-        mlflow.evaluate(
-            model_info.model_uri,
-            data,
-            extra_metrics=[mlflow.metrics.latency()],
-            evaluators="default",
-            evaluator_config={
-                "eval_results_path": "my_path",
-                "eval_results_mode": "overwrite",
-            },
-        )
-
-        mlflow.evaluate(
-            model_info.model_uri,
-            data,
-            extra_metrics=[mlflow.metrics.latency()],
-            evaluators="default",
-            evaluator_config={
-                "eval_results_path": "my_path",
-                "eval_results_mode": "append",
-            },
-        )
-
-        eval_table_from_delta = spark_session.read.format("delta").load(f"{tmpdir}/my_path")
-
-        assert eval_table_from_delta.count() == 4
